Add cross-chain sync and claim modules to also display when wallet is not connected
